fix: move dock icon setting into App::new after Iced initializes
Setting the dock icon before iced::application().run() had no effect
because Iced resets NSApplication state during its own initialization.
Moving the call into App::new() ensures it runs after Iced has set up
the application.
